Track unused elements and cleanup at end
There is a bug, as nicely documented with example in:

https://github.com/vugu/vugu/issues/101

Where vg-for ranges do not appropriately delete elements when the range
shrinks.

This happens because the DOM rendering re-uses the parent node, and
re-uses a subset of the existing child elements, but does not delete
the remaining child elements.

This commit adds a field which tracks the children of any re-used
elements, deleting them as they themselves are re-used, and finally
removes those which were never re-used at the end of the render.

This commit does include a new opcode, to be called to trigger cleaning
up the orphaned nodes.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>