Add buffered set inner html opcode
When a user tries to have the framework set a large HTML element (for
instance, large blocks of preformatted text), if the text itself exceeds
the buffer size minus some overhead (today, 16384 bytes), then the
application panics with 'requested instruction does not fit in the
buffer'.

This commit adds a new opcode opcodeBufferInnerHTML which can be called
in succession to buffer a large block of innerHTMl over multiple calls
on the JS side.  On the golang side, we loop over the innerhtml, so long
as it would overflow our buffer, we chunk up the inner html over
multiple calls to this opcode.  Finally, when the original setInnerHTML
op is performed, we prepend the buffer, if is not empty.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>