Better dynamic attributes
dynamic attributes may now contain one of:
- string
- any number type
- fmt.Stringer
- bool
- pointers

If a boolean type is given, the attribute will skipped if the value is
false. Else it will result in `attribute="attribute"`. This is usefull
for HTML flags like `disabled`.

In addition, `:="someObject"`, where `someObject` implements
`vugu.VGAttributeLister`, could be used to provide a dynamic list
of attributes.

implements #121

Signed-off-by: Thomas Berger <loki@lokis-chaos.de>