Test against Go 1.16 and 1.17
From the release notes for Go 1.16[1]:

    "Build commands like go build and go test no longer modify go.mod
     and go.sum by default. Instead, they report an error if a module
     requirement or checksum needs to be added or updated (as if the
     -mod=readonly flag were used). Module requirements and sums may
     be adjusted with go mod tidy or go get."

To accommodate this change we either need to create fully populated
`go.mod` and `go.sum` or call `go mod tidy` to satisfy `go build`.

`go mod tidy` seemed (by far) preferable, so now we call that at
strategic places in the tests as well as in the dev server handler.

[1]: https://go.dev/doc/go1.16