Use a Taskfile to build and test vugu
Use a Taskfile (see: https://taskfile.dev/) to build the vugu packages,
the vugugen and vugufmt command line tools, run all of the tests, and
lint the source code.

This simplifies building vugu locally. All that is required now is

$ task all

at the top level of the repo. The Taskfile.yaml is fully documented.

Note: this is an initial version that only supports Linux as it relies
on command line utilities that are installed by default on Linux. In
addition it assumes that docker, tinygo and the golangci-lint tools are
all installed locally. If they are not found then the task will abort.
Supporting Windows will require changing the tasks to support the
Windows PowerShell equivalents when the task in run on Windows.