Stop marking "main.wasm_gen.go" as an autogenerated file
Stop marking "main_wasm_gen.go" as an autogenerated file and rename it
back to "main_wasm.go".

At the minute it is not clear if this file should ever be generated, or
if a skeleton version should be generated by an external tool.
See the discussion around this in PR 261

https://github.com/vugu/vugu/pull/261