Run the github action workflow on any branch prefexed with " PR/"
In order to run the github action while developing a PR in a branch in a
forked respository we need to tell Gtihub the name of the branch.

But, we don't know until the PR request is made what the PR number will be so
we can't use the PR number as the branch name. We must use another
approach.

Github can trigger a build based on a pattern match of a brnach name,

See: https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#onpushbranchestagsbranches-ignoretags-ignore

So we pick a pattern of "PR/**" i.e. the branch name must begin with the
prefix "PR/".

This means that any branch named "PR/<something>" will trigger the
github action based build, regardless of which repository, a fork or the
upstream, the change set was pushed to.

This allows changes to github action itself to be developed as tested before the PR is made to the upstream respository.