Update the defaut tinygo docker image name to always use the lastest release
Update the hardcoded DefaultTinygoDockerImage value so that is always
references the lastest relase made by the TinyGo team.

Setting this to use the latest released tinygo container avoids build failures
where the version of tinygo in the container does not match the locally
installed version of Go. This is required becuase becuase TinyGo needs
access to the Go standard library and so the versions must align.

This isn't ideal, but is the simplest solution in the short term.

Long term, we need to make it clear that any docker image is downlaoded
and used at the level of the build scripts as opposed to the level of
the test code as is currently the case.