Update the task file so that it downlaods tinygo and golangci-lint
The task file has been updated so that is now downlaods and installs
the latest versions of both tinygo and the golangci-linter locally.

If these are already installed then they are upgraded to the latest
versions. If they are already at the latest versions then nothing is
changed.

This has been changed as we intend to remove this functionality fron the
GitHub action workflow and change the workflow so that it calls the the
task file.

This will result in a near identical experience between building vugu
locally and via the Github Action for maintainers and developers.

To achieve this the Taskfile calls out to a pair of external bash
scripts.

This is not ideal, as it relies not only on bash, but a Debian based
Linux install to work,

However this matches the CI/CD setup which is based on a ubuntu image.

Any cross platform update, including non-Debian based linux will need to
update or replace these scripts.