Use docker for both tinygo and golangci-lint and disable local tinygo testing
The taskfile has been reworked to use dockerized versions of both
tinygo and golangci-lint. This removes the need for the the locally
installed copies as well as the the external "download_*"
bash shell scripts which are OS specific.

This shows that the Taskfile can be simplified by using docker
as the package manager across OS's, at the expense of a slower lint time
- currently circa 40s.

This change also necessates the disabling of the tests that use a locally
installed tinygo compiler, which is now no longer installed.
The tests are currently commented out rather than removed.

Removal of the tests that used a locally installed copy of tinygo,
significantly speeds to the build. So overall the build time has been
reduced.