Remove the last test that was using a locally installed tinygo
Test024EventBufferSize wasn't following the naming or sequence
convention when running the test under standard go, a localy tinygo or
a dockererized tinygo.

As a result this test was missed when the local tinygo tests where
disabled. The test has now been disabled.

The naming has also been updated to reflect the covention used in the
other tests.

The Taskfile has also had a redundant comment removed.