Attempt to build test-001-simple outside of the wasm-test-suite
Attempt to build, and run the test test-001-simple outside of the
`wasm-test-suite` directory tree.

The mage based build takes a very different approach to achieveing this
compared to the existing `Taskfile` based approach. The previous approch
called the Go (and tinygo) compilers from within the test itself,
including building bespoke docker containers containing a Go web server
that was used to serve the wasm.

The new approach separates these concerns. The building of the
infrastructure, the building of the wasm to be served, and the
(building) and running of the test are all now separate steps taken at
the buildfile i.e Magefile level, rather than in the Go test files.

The new approach works as follows:

1) An nginx contianer is pulled.
2) A chromedp contianer is pulled.
3) A new docker bridge network is created called vugu-net.
4) The nginx contianer is started and connected to the vugu-net network.
5) The chromedp container is started and connected to the vugu-net
   network.
6) Both containers are also connected to the existing bridge network
   that allows outbound internet access from the host.
7) Run go generate in the tests directory (./new-tests/tetst-001-simple)
8) Build the wasm file for the test case we want to serve via nginx.
   nginx has already mounted the directory ./new-tests/test-001-simple
as the directory to /usr.share/nginx/html directory in the container.
9) Build and execute the test as normal i.e. go test
   github.com/vugu/vugu/new-tests/test-001-simple
10) The test will srart, connect to the chromedp container and have it
    requrest the wasm from the nginx container.
11) The chromedp container then runs the wasm file to produce the
    resulting DOM (HTML) page
12) The test then queries the DOM to check that the result is as
    expected. i.e. the wasm did what it was supposed to.
13) Then everything is torn down.

The commit includes an updated Magefile and Github action workflow file.