Update the Test001SimpleRefactor target with some debug traces
Added a pair of debug traces to the Test001SimpleRefactor target.
The targe now prints the current working directory and the listing of
that directory. It does this twice once using Go and once using shell in
case they differ on the github action runner.

This is an attempt to debug the go mod tidy issue showing up on the
runner but not locally.

See:
https://github.com/owenwaller/vugu/actions/runs/8911902154/job/24474195746

for an example