Fix the TestTinyGoCompiler test failure on MacOS
In TinygoCompiler.SetBuildDir the outpath is a temp directory.
That temp directory name is the full path, so it begins with a slash.
This makes the additional leading slash when appending the outpath to
"/src" to form the -o arguemnt is redundant. It is inserting a double
slash if the temp dir is "/tmp/dir" then the -o arguement is expanding to
"-o /src//tmp/dir" when we want "-o /src/tmp/dir"

On Linux this appears to be fine, but on MacOS it fails.
In either case it isn't correct.

The diff of the change is:

-               args = append(args, "-o", "/src/"+outpath)
+               args = append(args, "-o", "/src"+outpath)