Update the Test001SimpleRefactor target to strip the test name
Update the code for the Test001SimpleRefactor target such that the
directory name that contains the test does not have to be specified.

When building and running the test we use the 'go list -m' to deterine
the module name, the module name and the package name of the test are
currently idential, so either will do. Strictly both shoud be changed to
use the package name.

To determine which directory to mount into the nginx container we simply
use os.ReadDir. As there is only one test at the moment this is
sufficient.

A further change will be required to the test itself as well as
the magefile so that any test can be served.

The name of the directory (new-tests) that contains the test remaines hard
coded in the magefile at the minute.