Add WasmTestSuiteRefactor to support multiple tests
Refactored the previous Test001SimpleRefactor target to
WasmTestSuiteRefactor and refactored the test Test-001-Simple to support
being run as part of a suite of tests.

The magefile target WasmTestSuiteRefactor now passes the top level
tests directory - currently './new-tests' to the nginx container as a
the HTML root via a docker mount.

This means that all URL's must be prefixed with the package name. For
example

http://vugu-nginx/main.wasm

Becomes

http://vugu-nginx/test-001-simple/main.wasm

The test test-001-simple now generates the index.html file that is
served by the nginx container, from the index.html.tmpl file. This uses
the Go html template package to pass in the package name so that all of
the (local) URL's are correctly prefixed.

The test will read the template file, populate it, and write the
resulting index.html to disk.

Any pre-exisitng index.html file will be removed by the test prior to
generatign the new copy.

Test-001-simple has also been split into a root.vugu file containing the
HTML, a root.go file contianing the Go code.