Bump chromedp dependency to v0.9.5 from v0.9.3 for test-002 and test-003
Update the chromedp dependency from v0.9.3 to v0.9.5 for tests test-002-click
and test-003-prop.

This should prevent a build failure on Github actions in which a go
generate fails because the go.mod is outdated. This is true becuase the
go.mod under VCS has not been updated after the go generate has run.

To update the go.mod of a test several steps are required:

1) Delete the generated 0_components_gen.go file
2) delete the go.sum
3) remove all the "require" directives from the go.mod
4) run go generate to regenerate the 0_componets_gen.go file
5) run go mod tidy to update all the dependecies after the
   0_components_gen.go files has been generated.

The go.mod and go.sum should now be different,and show the updates.