Add an incomplete version of wasm test test-012-router
This is a refactred version of test-012-router from the wasm test suite.
While this version passes it is incomplete, and will have to be revised.

The wasm file of the test is served via the nginx container, that
serves from the parent 'new-tests' directory. As a result the vugu
router no longer appears to be working.

Uncommenting any of the lines in the test cases will result in a timeout
when the test is run.

The timeout occurs on both the path and the fragment version of the
test.