FIXME: Added a working but partial test-023-lifecylce-callbacks
This is a refactored version of test-023-lifecycle-callbacks from the
wasm-test-suite.

It is incomplete and needs to be fixed. The current test demonatrates
the lifecycle in action but appears to log only on the first pass though
the lifecycle.

This is different from the orginal wasm-test-suite test
in two respects.

Firstly, the revised test does not use a package level init fucntion.
This is problomatic as the init function is compiled for both wasm and
native code. As a result the call to js.Global().Call(...) will fail
with "js not implmented". Its unclear why this is happening.

Secondly, the log is now an propery of the Root object, rather then a
property - created via inline js - of the window object of the DOM.

The original wasm test suite verison of the code does pass, so its not
clear why the js.Global().Call() failure occurs in the revised test, or
why the test itself fails - logically the is the same in both cases.