FIXME: Add a comment to test-012-router that describes why the test is incomplete
The refactored test test-012-router from the (original) wasm-test-suite is
currently incomplete. This commit adds a FIXME comment to the code at
the failure point.

The problem is that the hard coded HTML hrefs in the page1/page2.vugu work,
but the links that are generated via the vgrouter and are called in response
to the button clicks return a 404 (or timeout).

This can be seen by connecting to the nginx container locally with a web
browser.

The revised version addes a URL path prefix of "/test-012-router". It is
unclear if this prefix which cannot be removed, as it related to the
filesystem path that nginx is using to serve the wasm, is causing the
vgrouter to fail or not.

This test therefore needs to be corrected, as the original
wasm-test-suite vetrsion of test-012-router does pass.