Added the TestSingleWasmTest build target to run a single wasm test
Added the TestSingleWasmTest build target. This target builds and runs a
single wasm test form the test suite, starting and destroying any containers.

The target takes the module name of the test to run as its sole
paramater. The target is not run as part of the AllXxx targets. It is
intended to be run explicitly when required.

The test must be specified by its module name, as defined in the tests
go.mod file.

The test must also exist as a separate directory under the
./new-tests directory. The TestSingleWasmTest target will only search
under the ./new-tests directory.