Add the StartLocalNginx and StopLocalNginx targets
Added the pair of targets StartLocalNginx and StopLocalNginx. These
start and stop a locally running nginx container that serves files from the
./new-tests directory where the wasm tests are located.

The intent is to allow a local browser/cURL etc to connect to the
nginx web server for debugging purposes.

To connect to a test case the URL is of the form:

http://localhost:8888/<wasm-test-directory>

e.g.

http://localhost:8888/test-001-simple

to run the wasm for the test case test-001-simple.

Running the StartLocalNginx twice without calling StopLocalNginx is
safe. The previously running nginx contianer will be stopped before any
further containers are started. Ditto if any of the TestWasm targets are
run, any locally running nginx container will be stoped first,