Rename the magefile targets and add the target documentation.
Update the magefile by:

Redefining 'All` so that it no longer runs the legacy wasm test suite.
Instead it always runs the newly refactored wasm test suite.

The legacy wasm test suite is only run by explicitly calling the
'TestLegacyWasm' target or the 'AllWithLegacyWasm' target. Both of these
targets are new.

The `AllFast' target has been removed.

The `AllFastNoLint` target has been renamed to `AllNoLint`.

The 'Build' target no longer checks for the presense of the docker
command - it never needed to in the first place.

Added and updated the docuemtnation for each target as displayed via the
`mage -l` command.

The constants used in the magefile are now in a separate const.go file.