Add the Examples and SingleExample magefile targets.
Add the new Example and SingleExample magefile targets to support
building and serving the examples.

The approach mirrors the approach taken for the wasm-test-suite package.
Each example is now a Go module in its own subdirectory. The example is
built via mage and servered using a local nginx container that is serving
from the local examples directory. Users can then connect to the example
using a URL of the form:

http://localhost:8888/<name-of-example-directory

e.g. for the fetch and display example

http://localhost:8888/fetch-and-display

SingleExample is the same as the Example target except it builds the
named example module, istead of all examples.

Running either target multiple times is safe, Any running nginx
contianer will be stoped before a new one is started.

Neither target is run by default as part of an AllXxx target. The
example targets must be run explicitly.