Add a sanity check to the build to find uncommitted or outdated files
In order to catch contributor mistakes, the magefile now checkes that:

All of the generated *_gen.go files have been commited to the repo and
match what is in the repo. The magefile does thas by running vugugen
again.

The go.mod and go.sum files match what is commited to the repo and
correctly reflect the state of go.mod and go.sum after vugugen has been run.

If there is a difference as detected by the git diff-files command
then the build aborts. The error message indicates the directory where
the uncommited files are found and a git status indicating which files
are uncommitted.

This check occurs against the repo root (i.e. the root go.mod), the
vgfrom package, each of the wasm-test-suite tests and each of the
examples.

The test is also carried out when an individual test or example is run.