Update github.com/chromedp/chromedp in the root go.mod to v0.9.5
Update the dependency github.com/chromedp/chromedp to v0.9.5 from
v0.9.3 in the root of the github.com/vugu/vugu module.

This is required to connect to the latest (as of 19th July 2024) version
of the chromedp container without generating the error:

ERROR: could not unmarshal event: parse error: expected string near offset 583 of 'cookiePart...'

in each of the wasm test suite tests.