Major refactor the magefile to improve the clarity of the build process
This is a major refactor of the top level magefile and all of the
helper files to improve clarity and simplify the build proces so that it
is easier to follow.

The process of building and running every wasm-test-suite test now
takes a module based approach. This brings the proces of building all of
wasm tests and a single wasm test - which is specified by module name -
into line.

The process works as follows:

Any necessary container images are pulled and the containers started.
This provides the nginx used to serve the wasm and the chromedp
container used by the wasm-test-suite tests.

The 'wasm-test-suite' top level directory is scanned for a list of modules, each
of which reside in a subdirectory.

This gives the absolute path to the module.

The build process then changes directory into this absolute path of
themoduel directory and executes the following for each module in turn:

* vugugen
* go mod tidy
* go build
* go test

before moving onto the next module.

When a single test is requested, the list of modules is first filtered,
to find the requested module. The build process then changes directory
to the specific module directory, before the above four commands are
executed.

This is in contrast to the previous approach. Previously each command
e.g. vugugen was executed in each module directory in turn. The
directory scan was then repeated and the next command, in this case go
mod tidy was executed in each module directory in turn. This process
continued for the other commands. Clearly this approach leads to
excessive and unecessary directory scanning. The new approach ensures
only one scan takes place (plus one filter operation if required)

The examples take the same approach as the wasm test suite, with the
exception that the examples do not execute the go test step and as a
consequence only the nginx contianer is started.

Numerous other code simplifications have taken place and a substantial
amount of code has been deleted - see vugugencmd.go for an example of
this - as a result of the refactoring.

See also `mage -l` as some of the top level targets have changed names -
mostly to do with container control, and a new
CleanLegacyWasmGeneratedFiles has been added to clean up the legacy wasm
test suite after the TestLegacyWasm target has executed.