Add targets to update the dependencies of each module
Added 4 new targets that all update dependencies:

UpgradeAllDependencies - which calls the Update* targets below
UpgradeRootDependencies - which updates the dependencies of the root
vugu module "github.com/vugu/vugu"
UpgradeWasmTestSuiteDependencies - which updates each of the wasm test
suite modules
UpgradeExampleDependencies - which updates each of the example modules

The UpgradeAllDependencies is not currently called as part of an
All* target. This may change in the future.

This update is required because running `go mod tidy` as part of the
Github Action CI/CD build will always pull the latest dependencies. The
Github Action build starts with an empty module cache this behaviour
applies to all modules i.e. ONLY the latest modules are ever used.

This behaviour would then cause the Github Action build to fail, because
the (newer) dependencies in the go.mod and go.sum did not match the
depenencies listed in the go.mod and go.sum in the repo.

Should this happen again in the future the solution is to run "mage
UpdateAllDependencies" if that target has not run automatically.