Add a new target SingleWasmTest to build and serve a single wasm test
Added a pair of new targets:

SingleWasmTest and SingleWasmTestWithGeneratedFileCheck

Both targets start the named wasm-test-suite module, and serve it via
a local nginx instance.

For example:

mage SingleExample gtihub.com/vugu/vugu/wasm-test-suite/test-002-click

would be served at:

http://localhost:8888/test-0020-click

This is useful if the behaviour of a test or the structure of the HTML
has to be examined in a browser. This can be useful in debugging the
tests.

The SingleWasmTestWithGeneratedFilesCheck target is the same but adds
the generated files check.