Add support for building with the last two supported versions of Go
Added support for building vugu with the last two supported versions of Go.

The Go team only support version v1.X and v1.(X-1) at the latest patch
level for each release. The go version in use is echoed to the Github
Aciton log.

For example if the currnet release was go v1.22.6 then the previous
release is 1.21.13. These are the Go versions that will be used to build
vugu.

The commit also removes the instalation the taskfile program, as this is no
longer used.