Split the workflow file into two, one each for each supported Go releases
It appears that the builds using the oldstable and stable versions of Go
are not isolated form each other. It appears that they are both running
on the same host/container/vm. As a result host wide options taken by
Docker are impacting both builds.

In an attempt to resolve this the workflow file has been split into two,
one for the oldstable release of Go and one for the current stable
release.