Rename each Github Action job to make the Go version clear
The name of each job in each of the YAML files has been updated to
reflect the Go version in use - stable or oldstable. The job was
previously called "Go tests".

This makes it easier to spot which job has failed in the Github Action UI as
the commit message and hash will be identical.