Re-enable linting via golangci-lint
Using a golangci-lint directive of the form:

//nolint:staticcheck // reason why staticcheck has been disabled

We can disable the use of the staticcheck linter completely at the blocks
or lines where is is currently reporting the use of the deprecated
ast.Package stuct.

This approach is documented here:
https://golangci-lint.run/usage/false-positives/#nolint-directive

At present golangci-lint will ignore staticheck //lint:ignore style
directives. This issues has been raised with the golangci-lint team. See:

https://github.com/golangci/golangci-lint/issues/2671#issuecomment-2310658099
and here:
https://github.com/golangci/golangci-lint/issues/1658#issuecomment-2312129214