Ensure that wasm_exec.js file is always copied from the GOROOT
The Go Wiki states:

Note: The same major Go version of the compiler and wasm_exec.js support file must be used together. That is, if main.wasm file is compiled using Go version 1.N, the corresponding wasm_exec.js file must also be copied from Go version 1.N. Other combinations are not supported.

See:
https://go.dev/wiki/WebAssembly

For the standard Go compiler ONLY this commit ensures that the above is
true for modules under both the wasm-test-suite and examples
directories.

It will also print a warning and a diff if a local repo
copy of wasm_exec.js differs from the GOROOT version.

The warning states that the GOROOT version will be used i.e. the local
copy will be overwritten. This is to ensure that the correct version of
wasm_exec.js for the compiler version in use is used. The warning will
not cause the built to fail.

The wasm_exec.js file does not contian any version information (e.g. as
a comment) so it is impossible to also say if any repo version is older
or newer than the Go compiler version in use. Regalrdss of this, the file
must be overwritten with the correct compiler version.

Once/If any local wasm_exec.js files are removed from the repo, the
check that produces the warning will have no effect. The change is
therefore forward compatible.

A separate change will be required to support the tinygo compiler at a
later date. The tinygo copy of wasm_exec.js is very different from the
standard compiler version and is located at a diffeent path under the
tinygo root. However, the same versioning constrains apply.