Update the vg-if example and add the Readme with the build instructions
Updated the `vg-if` example to add comments and tidy the code. The code
will now correctly print either "List is not empty, but hidden." or
"List is empty" but never both as was the case when the list was hidden
and empty.

There is now a Readme file that descrobes how to build and run the
example. This comes down to:

1) Clone the vugu repo
2) Run mage SingleExample github.com/vugu/vugu/examples/vg-if

The example is not go get'able. This is becuase `go get` only works in
module mode (which we need anyway). So cannot be used to download just
the source code of an example, from the `examples` package in the larger
`github.com/vugu/vugu` module.

A git clone will of course clone the wholw vugu repository, including
the magefiles. This is the only way to get the source code of the
examples.

At that point it's easier to just use mage to run the
examples.