Force the install of vugufmt to run in module-aware mode
A go install command can be put in to module-aware mode by specifiying
a verison to install e.g. @latest. We need to do this when we install the
vugufmt command so as not to change the root go.mod fiel of vugu itself

Putting go install in module-aware mode will cause go install to ignore
any local (or one found in a parent directory) go.mod file. This avoids
any attempt to change the local go.mod which may belong to a different
module.

See: https://go.dev/ref/mod#mod-commands

Since Go 1.16, if the arguments have version suffixes (like @latest or @v1.0.0), go install builds packages in module-aware mode, ignoring the go.mod file in the current directory or any parent directory if there is one. This is useful for installing executables without affecting the dependencies of the main module.

In this case 'vugu' itself would be considered the main module i.e. the
root go.mod file is the one at the root of the repository.