Optionally install the goimports via the magefile and not via Github Action
Update the magefiel to add a new InstallGoImports target.

This target firsstr checks if any version of gimports can be found on
the path. If not then it installs the latest version of the latest
vetsoin of the goimports command via:

go install golang.org/x/tools/cmd/goimports@latest

The @latest ensures that the dependencies listed in the root go.mod of
vugu are left unchanged.