An translation example that uses the embed package.
A very simple example that shows how a message on a webpage can be translated
into multiple languages.

The user sets their language preference via a
set of radio buttons on the web page. The message is then
translated to the seclected language.

The translation is based on a set of language files, one per supported
language. These files are embedded within the WASM binary using the
standard go embed package.

The translation is provided via the go-i18n package, which loads the
language files from the WASM binary prior to translation.