A simple example that shows how to use vugu components
A simple example that shows how to use vugu components. Each of the
three components are defined in their own *.vugu and *.go files.

The example consists of a Parent component and two button components.
The parent.vugu file demonstrates how to create the two buttons using
<pkg:Comp ...> syntax.
Clicking on either button component changes the message displayed in the
parent component.

This is achieved by setting a "Parent" field dynamcially in each button
when it is created using the ":FiledName" syntax. Each button then calls
the Parents SetMsg(string) method to change the message.