Revert "Oops- remove the duplicate and out dated Readme."
This reverts commit 63286ab07d5c25af547692dca3a10e1e0704e5cb.

Opps, commit 63286ab accidently removes the entire README.md file, and not
a duplicate of it. This commit reverses the deletion  and restores the file.

Without a README.md file, Github has been unable to display any project
information to visitors to the vugu repo. With this commit Github will
be abe to display the project inforation once again.