An HTML form example that demonstrates how to use vugu with HTML forms
The example demonstrates how to use vugu to build and process an HTML form
containing a number of sub controls.

The example does not use the vgform package in vugu. This example is
built with the raw HTML controls and vugu itself.

The form contains:

a single line text input control
a single line email address input control
a set of radio buttons
a set of statically set options in a drop down list
a set of dynamically generated options in a drop down list.

The example demonstrates how to process the values of these controls as
a single form based on a single "Submit" button.

When the "Submit" button is pressed the form data is read and converted
to a JSON structure. This structure is then displayed on the web page.

The example does not sent the form data to a remote server. To send the
data the Sampleform.Submit(..) method would need to be updated to make
an HTTP Post call to the remote server with the JSON encoded data. This
is left as an exercise for the reader.

The example also splits the form into a number of (vugu)
(sub)components. So for example the email address input control is a
stand alone control, see Emailinput.go and Emailinput.vugu.

Each vugu (sub)component is passed a reference to the
controlling(enclosing)
Sampleform component dynamically at initialisation time. See
Sampleform.vugu for the details.