Rework the radio buttons group to better support setting the default option
The radio buttons have been reworked to make it easier to set the
default (or initially checked value) in the set of radio buttons.

This removes any duplication in the setting of the default value
between the vugu and Go sides. The default value is now wholly set in
the vugu side when the radio group is constructed.

To achieve this a new Language Radio Button component is required. This
new component must have its HTML/JS properties set as fields in the
underlying Go struct.

The vugu file for the Language Radio Button will
then set the struct values to the HTML property values for the
underlying <input type="radio" ..> HTML element.

This is the only way for the Language Radio Button to determine if
it is also the default option, and show should display a checked radio
button at initialisation.

This is achieved by comparing the current
value (taken from the struct field) to the default value in the
containing radio group. If they are the same then the "checked"
attribute of the radio button HTML control will be set.

We must use the structure field value beacause
outside of an event there is no way to read the "value" property of a
"target".

Since this process must occur at component initialisation no event is
available.