Added test-025-valueof a simple test case for issue #328 - js.ValueOf bug
A very simple test case that shows how calling vugu.js.Valueof will
result in a panic (ValueOf: Invalid value).

The panic is caused because the underlying type passed into js.ValueOf in
the test is a vugu.js.Value.

vugu.js.ValueOf is a wrapper around syscall.js.ValueOf.

The implementation of syscall.js.valueOf is a type switch on the value
passed to it. Since a vugu.js.Value is unknown to syscall.js.ValueOf we
trigger the default case which results in a panic.

See: https://github.com/vugu/vugu/issues/328 for the discussion.