Add a static render test that calls into vugu/js
Added a static renderer test that makes a call into the vugu/js package.
The test will be built for and run on the native target.

This test confirms that vugu/js is as functional as we need i.e. it does not
panic on targets other than GOOS=js GOARCH=wasm.

This is test is related to Issues #330 and #328
https://github.com/vugu/vugu/issues/330
https://github.com/vugu/vugu/issues/328